requirements.txt
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
werkzeug==0.16.1
wheel==0.24.0
pyyaml==5.4.1
icecream
